:recycle: 산술연산자 테스트코드 구현 및 기존 코드 리팩토링
